requirements.txt
boto3>=1.34.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-aws>=0.1.0
pinecone-client>=3.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
tiktoken>=0.6.0
